fixing the backend stuctuer and engagment system startup creating a navbar component
