+++ backend/app/auth.py
@@ -3,7 +3,7 @@ from firebase_admin import auth, credentials
 from fastapi import HTTPException
 
 # Initialize Firebase Admin SDK
-cred = credentials.Certificate("path/to/serviceAccountKey.json")
+cred = credentials.Certificate("credentials/research-project-8003b-firebase-adminsdk-b4h3x-b4ea62e36f.json")
 firebase_admin.initialize_app(cred)
 
 # Function to verify the Firebase token
