+++ backend/credentials/research-project-8003b-firebase-adminsdk-b4h3x-b4ea62e36f.json
@@ -1,0 +1,13 @@
+{
+  "type": "service_account",
+  "project_id": "research-project-8003b",
+  "private_key_id": "b4ea62e36fbc198148c553a3766414b7512770b2",
+  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQCf/gJqT9Set7fn\n/3q8+OJInmFI/kBjdzzx0vNoCMG4htAKtWlSgInKPLy7Ok5Jcco3XjavgJGfDVxb\nGfMDV3qmBnktS2l9vD6YyaAiIs7RU/d8mFj8KBAKBKrTqBau+9y+nWtMPejobrUP\n0PrzUOsYLsUgcfiTL7dV1UwPWqV907Oe+emSeY/V8/1nzvWm1rw40IH5e6Gm0HDc\nU1O19LvR+l7PjhckpQRQQYkNSrSd5SmwqI1M7K0rPAo3Q/MWRvLEQnq3evOTuayI\ne/0y7CbP2JNnSJtGavEtUxo8XJDrY2a2u3R4+WYZLaD/ngQ/q4+ha/ckyNUJTeDz\nXoAcxXTBAgMBAAECggEAFPWy8qMuCJ5VjSLOvTj3Hr+uS7YQpgyLQxGkFP9TWaLt\nAv93Inl3vl3XSnPoqP3WRGCnALZIMt0OOgUb7VZ62TXvOWbawBUhjOSagj5VzTNB\n2znI1pKXtKhIijKdc3kdbH9fp3z+KTfE/U3Jwocu9Vu93sfvjQtq6jy8789WmjV6\nbY7jq9JWMtb/sG5ICVTm+LzdOIS3nMWbDx1BB9Vxg98FBqWBgaou3RU6KWXnKeCa\nFXeJTLjpcXIY3vKYBIqTdbASaG2KXHPl5PCH6xEFrBHpUCaINa5Mhb6SHr+IdmuE\nOGq5730HiyGh4A8pgSz9j578Q61pg8Vo5+wHOmJXLQKBgQDMQ3Tw1yc/g2/GHD5E\niSHLAKAvLdBHZ6Dk+LcqPJlznLWlsjyjPgOOYLZG7+t6zEqdVy1M+HvF8quRkYzF\nPbTN8Xl+fsXAz5v7AO5s14QBVbD35N0ZguncD4761RcePWmOEeQZAKJfzYg7sD1O\nH0scT1tYK+CNJN/8qVJnmFwEWwKBgQDIg/tUMIM5I6NW2Ul0aPJ8WNfqtmG66Mn+\ni1oVOcCev3cB/2A3tHe4h4JgGiFCUDHDFPF0EJc+PJBIboKaQc7DFbqFZS/DUdvi\n+ylF8gt2cBlQSfX6FFVQlxlkzuVrqf20Efk6oASFBgIEe69XoDpSe8100iebULAG\nnmbibTUGEwKBgHO6iGfnzpHamxpoaBMWivHkId15HHdITE1OaSBiAyMQl44QhPCM\nMyGYw80/I3CIjIHGoaDpa98EFZa9Gz+pLWY9OPOlSc9rENnOdkdud5j1/vUEUwyg\nMnV+zyhoUNfOPZ7DnjNYO6cB4PzIZoN7kWexwnQbL/dn/Au/tzOAjCnfAoGBAMVs\nHTKBKVJC4nmKKqZxBf6kQsKXm5W5ydfTQHiRijNC9FuFIyj1iMKouRcB4yk1aClZ\nH3pVM3RVkQGfvPtAO+EcLa9VNzeqC0NqYvK5I8wTBvlkOMlm57EXdCsrAfuFPYPk\nBm1QS+LlTWs4lCVEXugDeQDCQbjMMuPxZf8d4wz9AoGBAJ6d5Gsmsdfa6MM6L4OH\n+xPfjzquXNgdE4o6sKw6KFBXxXJMyXlrqHlWAstCL2oLaxaIdAwHUFyKgUuiBn9Y\n7XBmN1yGYUabL5g9SFsTYPa41UITAUL87bz77RKvyai631VOgs1nzl3FowTqP385\n3DzgOdH8j64rq84cGrZZ18m6\n-----END PRIVATE KEY-----\n",
+  "client_email": "firebase-adminsdk-b4h3x@research-project-8003b.iam.gserviceaccount.com",
+  "client_id": "111343861130140153736",
+  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+  "token_uri": "https://oauth2.googleapis.com/token",
+  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-b4h3x%40research-project-8003b.iam.gserviceaccount.com",
+  "universe_domain": "googleapis.com"
+}
