+++ src/components/Navbar.css
@@ -1,0 +1,49 @@
+/* Navbar.css */
+.navbar {
+    background-color: #343a40; /* Dark background color */
+    padding: 10px 20px;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1); /* Subtle shadow */
+  }
+  
+  .nav-list {
+    list-style-type: none;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin: 0;
+    padding: 0;
+  }
+  
+  .nav-item {
+    margin: 0 15px;
+  }
+  
+  .nav-link {
+    text-decoration: none;
+    color: #ffffff; /* White text color */
+    font-weight: bold;
+    padding: 8px 15px;
+    border-radius: 4px;
+    transition: background-color 0.3s, color 0.3s; /* Smooth transition */
+  }
+  
+  .nav-link:hover {
+    background-color: #495057; /* Darker shade on hover */
+    color: #ffffff; /* White text color */
+  }
+  
+  .sign-out-button {
+    background: none;
+    border: none;
+    color: #ffffff;
+    font-weight: bold;
+    cursor: pointer;
+    padding: 8px 15px;
+    border-radius: 4px;
+    transition: background-color 0.3s, color 0.3s;
+  }
+  
+  .sign-out-button:hover {
+    background-color: #dc3545; /* Red color on hover */
+  }
+
